# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  org.CreateTeamDTO:
    properties:
      avatar:
        maxLength: 255
        type: string
      description:
        maxLength: 500
        type: string
      display_name:
        maxLength: 100
        minLength: 2
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - display_name
    - name
    type: object
...
  org.TeamDTO:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      description:
        type: string
      display_name:
        type: string
      id:
        type: integer
      name:
        type: string
      org_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  response.DataResponse-org_TeamDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/org.TeamDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.DataResponse-twofa_EnableDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/twofa.EnableDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-twofa_SetupDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/twofa.SetupDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-twofa_StatusDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/twofa.StatusDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.ErrorDetail:
    properties:
      code:
        description: 业务错误码（小写下划线）
        type: string
      details:
        description: 额外详情（如验证错误列表）
      message:
        description: 错误消息
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
    type: object
  response.MessageResponse:
    description: 纯消息响应格式
    properties:
      code:
        description: HTTP 状态码
        type: integer
      message:
        description: 消息描述
        type: string
    type: object
...
  response.PagedResponse-org_TeamDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        description: 数据列表
        items:
          $ref: '#/definitions/org.TeamDTO'
        type: array
      message:
        description: 消息描述
        type: string
      meta:
        allOf:
        - $ref: '#/definitions/response.PaginationMeta'
        description: 分页信息
    type: object
...
  response.PaginationMeta:
    properties:
      has_more:
        description: 是否有下一页
        type: boolean
      page:
        description: 当前页码
        type: integer
      per_page:
        description: 每页数量
        type: integer
      total:
        description: 总记录数
        type: integer
      total_pages:
        description: 总页数
        type: integer
      warning:
        description: 页码越界警告
        type: string
    type: object
...
  twofa.EnableDTO:
    properties:
      message:
        description: 提示消息
        type: string
      recovery_codes:
        description: 恢复码列表
        items:
          type: string
        type: array
    type: object
  twofa.SetupDTO:
    properties:
      qrcode_img:
        description: Base64 编码的二维码图片
        type: string
      qrcode_url:
        description: 二维码 URL
        type: string
      secret:
        description: TOTP 密钥（用户可手动输入）
        type: string
    type: object
  twofa.StatusDTO:
    properties:
      enabled:
        description: 是否启用
        type: boolean
      recovery_codes_count:
        description: 剩余恢复码数量
        type: integer
    type: object
  twofa.VerifyDTO:
    properties:
      code:
        description: TOTP 验证码
        example: "123456"
        type: string
    required:
    - code
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/lwmacct/260103-ddd-iam-bc
  description: 基于 DDD + CQRS 架构的可复用模块库
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go DDD Package Library API
  version: "1.0"
paths:
...
  /api/auth/2fa/disable:
    post:
      consumes:
      - application/json
      description: 禁用当前用户的两步验证功能
      produces:
      - application/json
      responses:
        "200":
          description: 2FA禁用成功
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "400":
          description: 禁用失败
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 禁用 2FA
      tags:
      - auth-2fa
  /api/auth/2fa/setup:
    post:
      consumes:
      - application/json
      description: 为当前用户生成2FA密钥和二维码，用于配置身份验证器应用
      produces:
      - application/json
      responses:
        "200":
          description: 2FA初始化成功
          schema:
            $ref: '#/definitions/response.DataResponse-twofa_SetupDTO'
        "400":
          description: 设置失败
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 设置 2FA
      tags:
      - auth-2fa
  /api/auth/2fa/status:
    get:
      consumes:
      - application/json
      description: 获取当前用户的2FA启用状态和剩余恢复代码数量
      produces:
      - application/json
      responses:
        "200":
          description: 2FA状态
          schema:
            $ref: '#/definitions/response.DataResponse-twofa_StatusDTO'
        "400":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 2FA 状态
      tags:
      - auth-2fa
  /api/auth/2fa/verify:
    post:
      consumes:
      - application/json
      description: 验证身份验证器应用生成的TOTP代码，成功后启用2FA并返回恢复代码
      parameters:
      - description: TOTP验证码
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/twofa.VerifyDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 2FA启用成功
          schema:
            $ref: '#/definitions/response.DataResponse-twofa_EnableDTO'
        "400":
          description: 验证码错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 启用 2FA
      tags:
      - auth-2fa
...
  /api/org/{org_id}/teams:
    get:
      consumes:
      - application/json
      description: 分页获取组织内的团队列表
      parameters:
      - description: 组织ID
        in: path
        minimum: 1
        name: org_id
        required: true
        type: integer
      - default: 20
        description: Limit 每页数量，默认 20，最大 1000
        in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: Page 页码，从 1 开始
        in: query
        minimum: 1
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 团队列表
          schema:
            $ref: '#/definitions/response.PagedResponse-org_TeamDTO'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 团队列表
      tags:
      - org-team
    post:
      consumes:
      - application/json
      description: 在组织内创建新团队
      parameters:
      - description: 组织ID
        in: path
        minimum: 1
        name: org_id
        required: true
        type: integer
      - description: 团队信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/org.CreateTeamDTO'
      produces:
      - application/json
      responses:
        "201":
          description: 团队创建成功
          schema:
            $ref: '#/definitions/response.DataResponse-org_TeamDTO'
        "400":
          description: 参数错误或团队名已存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建团队
      tags:
      - org-team
  /api/org/{org_id}/teams/{team_id}:
    delete:
      consumes:
      - application/json
      description: 软删除团队
      parameters:
      - description: 组织ID
        in: path
        minimum: 1
        name: org_id
        required: true
        type: integer
      - description: 团队ID
        in: path
        minimum: 1
        name: team_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 团队删除成功
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "400":
          description: 无效的ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 团队不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除团队
      tags:
      - org-team
    get:
      consumes:
      - application/json
      description: 获取团队详情
      parameters:
      - description: 组织ID
        in: path
        minimum: 1
        name: org_id
        required: true
        type: integer
      - description: 团队ID
        in: path
        minimum: 1
        name: team_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 团队详情
          schema:
            $ref: '#/definitions/response.DataResponse-org_TeamDTO'
        "400":
          description: 无效的ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 团队不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 团队详情
      tags:
      - org-team
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"